% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_topo.R
\name{geom_topo}
\alias{geom_topo}
\title{Générer une carte topographique avec effets d’ombrage et lignes de contour}
\usage{
geom_topo(raster_elevation, intervalle = NULL, ombrage = TRUE)
}
\arguments{
\item{raster_elevation}{Un objet `SpatRaster` ou un chemin vers un fichier `.tif`}

\item{intervalle}{(optionnel) Intervalle pour les lignes de contour (en mètres)}

\item{ombrage}{Logique. Si TRUE, ajoute un effet d’ombrage}
}
\value{
Une liste de couches ggplot à ajouter à une carte
}
\description{
Cette fonction crée une carte topographique à partir d’un raster d’élévation,
avec la possibilité d’ajouter un effet d’ombrage (hillshade) et des lignes de
contours selon un intervalle spécifié.
}
