% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidite.R
\name{indice_humidite}
\alias{indice_humidite}
\title{Calculer les métriques topographiques d'humidité}
\usage{
indice_humidite(raster_mnt, dossier = NULL, prefixe = "")
}
\arguments{
\item{raster_mnt}{Un objet `SpatRaster` ou un chemin vers un fichier `.tif`}

\item{dossier}{Chemin du dossier de sortie (facultatif)}

\item{prefixe}{Préfixe pour les noms de fichiers exportés (si `dossier` est défini)}
}
\value{
Un objet `SpatRaster` avec les couches "Geomorphons", "TWI", "Depression" et "Accumulation"
}
\description{
Cette fonction calcule l'indice d'humidité topographique (TWI), la profondeur des dépressions,
l'accumulation de flux et les formes du terrain (géomorphons) à partir d'un MNT ou MNE.
}
