% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar.R
\name{lidar}
\alias{lidar}
\title{Télécharger et recadrer les données LiDAR pour une zone donnée}
\usage{
lidar(polygone, dossier = NULL, mne = FALSE, recent = FALSE)
}
\arguments{
\item{polygone}{Un objet `sf` représentant la zone d’intérêt ou un chemin vers un fichier vectoriel (`.shp`, `.gpkg`, etc.)}

\item{dossier}{Dossier de sortie pour enregistrer les fichiers raster (optionnel)}

\item{mne}{Logique. Si TRUE, télécharge le MNE (modèle de surface). Sinon, le MNT.}

\item{recent}{Logique. Si TRUE, conserve uniquement la version la plus récente}
}
\value{
Une liste de `SpatRaster` (un par année)
}
\description{
Cette fonction interroge l'API STAC de RNCan pour télécharger les données LiDAR (MNE ou MNT)
correspondant à une géométrie fournie, puis les recadre à cette zone d’intérêt.
}
