% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplane_sf.R
\name{multiplane_sf}
\alias{multiplane_sf}
\title{Convertir un fichier MultiPlane en objet sf}
\usage{
multiplane_sf(fichier, imperial = TRUE)
}
\arguments{
\item{fichier}{Chemin vers le fichier texte MultiPlane (tabulations)}

\item{imperial}{Logique. Si TRUE, les distances sont en pieds (sinon en mètres)}
}
\value{
Un objet `sf` avec les colonnes `elevation`, `code` et `geometry`
}
\description{
Cette fonction lit un fichier texte MultiPlane (issu d’un levé GPS ou station totale) et le convertit
en un objet `sf` avec projection WGS84. Les coordonnées sont calculées à partir d’un point de référence.
}
