% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raccourcir_lignes.R
\name{raccourcir_lignes}
\alias{raccourcir_lignes}
\title{Raccourcir des lignes en interpolant les extrémités}
\usage{
raccourcir_lignes(lignes, distance, extremite = c("debut", "fin"))
}
\arguments{
\item{lignes}{Objet `sf` contenant des géométries `LINESTRING`}

\item{distance}{Distance à retirer (en unités du système de coordonnées)}

\item{extremite}{Choix de l’extrémité à couper : "debut", "fin" ou `c("debut", "fin")`}
}
\value{
Un objet `sf` avec les lignes raccourcies
}
\description{
Cette fonction permet de raccourcir une ou plusieurs lignes `LINESTRING` en retirant une distance
au début, à la fin, ou aux deux extrémités. Elle conserve une géométrie valide en interpolant les points.
}
