% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reseau_ligne_parallele.R
\name{reseau_lignes_paralleles}
\alias{reseau_lignes_paralleles}
\title{Créer un réseau de lignes parallèles orientées avec angle et coin dynamiques}
\usage{
reseau_lignes_paralleles(polygone, angle = NULL, largeur = 20, coin = "SO")
}
\arguments{
\item{polygone}{Objet `sf` (POLYGON ou MULTIPOLYGON)}

\item{angle}{(optionnel) Angle en degrés. Si NULL, calculé automatiquement.}

\item{largeur}{Espacement entre les lignes.}

\item{coin}{Coin d’ancrage : `"SO"`, `"SE"`, `"NE"`, `"NO"` (par défaut `"SO"`)}
}
\value{
Un objet `sf` de lignes coupées à l'emprise du polygone
}
\description{
Cette fonction génère un réseau de lignes parallèles orientées selon un angle donné ou calculé automatiquement,
espacées d'une largeur fixe, et couvrant une emprise fournie. Permet de choisir le coin de départ.
}
