% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinel2.R
\name{sentinel2}
\alias{sentinel2}
\title{Télécharger et calculer un indice de végétation Sentinel-2 pour une parcelle}
\usage{
sentinel2(
  polygone,
  dossier = NULL,
  annee_debut = 2018,
  annee_fin = NULL,
  max_nuage = 50,
  mois = 3:11,
  nuage_dans_la_parcelle = 5,
  indice = "NDVI"
)
}
\arguments{
\item{polygone}{Un objet `sf` ou un chemin vers un fichier vectoriel (shapefile, geojson, etc.)}

\item{dossier}{(optionnel) Dossier dans lequel enregistrer les fichiers raster (.tif)}

\item{annee_debut}{Année de début (par défaut `2018`)}

\item{annee_fin}{Année de fin (par défaut : année courante)}

\item{max_nuage}{Pourcentage maximal de nuages (0 à 100) selon les métadonnées}

\item{mois}{Mois d'intérêt (entiers entre 1 et 12)}

\item{nuage_dans_la_parcelle}{Proportion maximale de nuages dans la parcelle (0–100), ou `FALSE` pour désactiver}

\item{indice}{Indice de végétation à calculer (`"NDVI"`, `"GNDVI"`, `"EVI"`, `"KNDVI"`, , `"NDRE"`)}
}
\value{
Un objet `SpatRaster` contenant l'indice de végétation par date
}
\description{
Cette fonction interroge le Planetary Computer pour obtenir des images Sentinel-2 L2A,
filtre selon la couverture nuageuse, calcule un indice de végétation (NDVI, GNDVI, EVI, KNDVI, NDRE)
et retourne une stack raster avec les valeurs de l'indice sélectionné.
}
