% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_ags.R
\name{sf_ags}
\alias{sf_ags}
\title{Écrire un fichier AGS à partir d'un objet sf}
\usage{
sf_ags(
  sf_obj,
  fichier_sortie,
  colonnes = list(lat = "*Lattitude (deg)", lon = " Longitude (deg)", elev =
    "Elevation Existing (m)", info = "Code")
)
}
\arguments{
\item{sf_obj}{Un objet `sf` avec des géométries de points}

\item{fichier_sortie}{Chemin de base pour le fichier de sortie (sans extension)}

\item{colonnes}{Liste nommée : noms internes des colonnes à écrire = noms d'en-tête souhaités
Exemple : list(lat = "*Lattitude (deg)", lon = " Longitude (deg)", elev = "Elevation Existing (m)", info = "Code")}
}
\value{
Rien. Écrit un fichier `.ags`.
}
\description{
Écrire un fichier AGS à partir d'un objet sf
}
