% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_multiplane.R
\name{sf_multiplane}
\alias{sf_multiplane}
\title{Convertir des points sf en fichier texte MultiPlane}
\usage{
sf_multiplane(
  points,
  benchmark,
  colonne_z = "Z",
  colonne_code = "Code",
  fichier_sortie,
  altitude_benchmark = 0,
  imperial = FALSE
)
}
\arguments{
\item{points}{Objet `sf` avec les points à convertir}

\item{benchmark}{Objet `sf` contenant un seul point de référence}

\item{colonne_z}{Nom de la colonne pour l’altitude}

\item{colonne_code}{Nom de la colonne indiquant les codes de contours}

\item{fichier_sortie}{Chemin du fichier `.txt` à générer}

\item{altitude_benchmark}{Altitude du point de référence (par défaut : 0)}

\item{imperial}{Logique. Si TRUE, convertit les unités en pieds}
}
\value{
Un `tibble` contenant les données formatées pour MultiPlane
}
\description{
Cette fonction convertit un objet `sf` contenant des points en format tabulé compatible avec
MultiPlane, à partir d’un point de référence (benchmark). Les distances X/Y et l’altitude sont calculées
par rapport au benchmark.
}
