% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grille.R
\name{st_grille}
\alias{st_grille}
\title{Générer une grille orientée selon un polygone}
\usage{
st_grille(
  polygone,
  angle = NULL,
  largeur = 100,
  longueur = 100,
  coin = "SO",
  quinconce = FALSE,
  decouper = TRUE,
  points = FALSE
)
}
\arguments{
\item{polygone}{Objet `sf` de type POLYGON ou MULTIPOLYGON}

\item{angle}{Angle d’orientation en degrés. Si NULL, il est déduit automatiquement.}

\item{largeur}{Largeur des cellules (axe court)}

\item{longueur}{Longueur des cellules (axe long, optionnel)}

\item{coin}{Coin de départ : "SO", "SE", "NE" ou "NO"}

\item{quinconce}{Logique : grille en quinconce (TRUE/FALSE)}

\item{decouper}{Logique : découper les cellules au bord du polygone}

\item{points}{Logique : retourner les centroïdes au lieu des polygones}
}
\value{
Un objet `sf` avec les colonnes `ligne`, `colonne`, et `geometry`
}
\description{
Cette fonction crée une grille de polygones alignée à l'enveloppe orientée minimale d’un polygone.
Elle permet d’indiquer un angle, une largeur, une longueur, un coin d’origine, un décalage en quinconce,
et de restreindre les cellules à celles qui croisent ou sont contenues dans le polygone.
}
