% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_utm.R
\name{st_utm}
\alias{st_utm}
\title{Reprojeter un objet sf dans la bonne zone UTM}
\usage{
st_utm(x)
}
\arguments{
\item{x}{Un objet `sf` à reprojeter}
}
\value{
Un objet `sf` reprojeté dans la zone UTM correspondante
}
\description{
Cette fonction détecte automatiquement la zone UTM appropriée en fonction du centroïde d’un objet `sf`,
puis transforme la géométrie vers ce système de coordonnées métriques (UTM).
}
