% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_sf.R
\name{svy_sf}
\alias{svy_sf}
\title{Charger un fichier SVY et retourner un objet sf}
\usage{
svy_sf(fichier)
}
\arguments{
\item{fichier}{Chemin vers un fichier .SVY}
}
\value{
Un objet `sf` avec les colonnes lat, lon, elev et info
}
\description{
Charger un fichier SVY et retourner un objet sf
}
