% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone.R
\name{zones}
\alias{zones}
\title{Générer des zones de gestion à partir de superpixels}
\usage{
zones(s_poly, w, df, n_zones = 5, superficie_min = 10000)
}
\arguments{
\item{s_poly}{Objet `sf` contenant les superpixels}

\item{w}{Objet de poids spatiaux (issu de `rgeoda::knn_weights` ou équivalent)}

\item{df}{Tableau de données associé, sans géométrie, incluant la colonne `area`}

\item{n_zones}{Nombre de zones de gestion souhaitées}

\item{superficie_min}{Surface minimale par zone (en mètres carrés)}
}
\value{
Un objet `sf` représentant les zones fusionnées par groupe
}
\description{
Cette fonction applique l’algorithme SKATER sur des superpixels et leurs poids spatiaux,
afin de générer des zones de gestion respectant une contrainte de superficie minimale.
}
