% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reseau_ligne_parallele.R
\name{reseau_lignes_paralleles}
\alias{reseau_lignes_paralleles}
\title{Créer un réseau de lignes parallèles orientées}
\usage{
reseau_lignes_paralleles(polygone, angle, largeur)
}
\arguments{
\item{polygone}{Un objet `sf` de type POLYGON, MULTIPOLYGON ou `bbox` représentant la zone à couvrir}

\item{angle}{Angle d’orientation des lignes en degrés (par rapport à l’axe des x, dans le sens antihoraire)}

\item{largeur}{Espacement entre les lignes (en unités du système de coordonnées)}
}
\value{
Un objet `sf` de type LINESTRING contenant le réseau de lignes découpé à l'emprise
}
\description{
Cette fonction génère un réseau de lignes parallèles orientées selon un angle donné,
espacées d'une largeur fixe, et couvrant une emprise fournie.
}
